refactor: Make CharacterItem class
