+++ feature/home/src/main/java/com/teamoffroad/feature/home/HomeScreen.kt
@@ -21,13 +21,11 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.teamoffroad.core.designsystem.theme.Main1
 import com.teamoffroad.core.designsystem.theme.OffroadTheme
-import com.teamoffroad.feature.home.presentation.component.character.CharacterImage
-import com.teamoffroad.feature.home.presentation.component.character.CharacterNameText
-import com.teamoffroad.feature.home.presentation.component.quest.progressbar.CloseCompleteRequest
-import com.teamoffroad.feature.home.presentation.component.character.EmblemNameText
 import com.teamoffroad.feature.home.presentation.component.HomeIcons
-import com.teamoffroad.feature.home.presentation.component.user.NicknameText
+import com.teamoffroad.feature.home.presentation.component.character.CharacterItem
+import com.teamoffroad.feature.home.presentation.component.quest.progressbar.CloseCompleteRequest
 import com.teamoffroad.feature.home.presentation.component.quest.progressbar.RecentQuest
+import com.teamoffroad.feature.home.presentation.component.user.NicknameText
 import com.teamoffroad.feature.home.presentation.model.HomeProgressBarModel
 import com.teamoffroad.offroad.feature.home.R
 
@@ -56,7 +54,7 @@ private fun UsersAdventuresInformation(modifier: Modifier = Modifier) {
     ) {
         Column {
             NicknameText("비포장도로")
-            CharacterNameText("오푸")
+            CharacterItem().CharacterNameText("오푸")
         }
         Box {
             Box(
@@ -66,11 +64,11 @@ private fun UsersAdventuresInformation(modifier: Modifier = Modifier) {
                 HomeBackground()
                 HomeIcons()
             }
-            CharacterImage()
+            CharacterItem().CharacterImage()
         }
     }
     Spacer(modifier = Modifier.padding(18.dp))
-    EmblemNameText(Modifier)
+    CharacterItem().EmblemNameText(Modifier)
 }
 
 @Composable
