--- feature/home/src/main/java/com/teamoffroad/feature/home/presentation/component/character/CharacterImage.kt
@@ -1,27 +1,0 @@
-package com.teamoffroad.feature.home.presentation.component.character
-
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.unit.dp
-import com.teamoffroad.offroad.feature.home.R
-
-@Composable
-fun CharacterImage() {
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(top = 90.dp),
-        contentAlignment = Alignment.Center
-    ) {
-        Image(
-            painter = painterResource(id = R.drawable.img_home_explorer),
-            contentDescription = "explorer"
-        )
-    }
-}
